Normative-Types-Specification: Remove some BNF
Sphinx stumbled across some lines having a single "/",
which probably stands for "more content to be added".

Remove those lines, togther with the surrounding text blocks.
While there, remove some of the definitions of our syntax,
which is a inspired by the extended Backus–Naur form (EBNF).

However, we don't follow that standard here fully anyway,
[] is used for arrays, and ":opt" is used for an optional field.

The idea is to give the reader an idea how structures can be
defined, and which definition are "good practice", loosing
the requirement for a fully-syntax-conform specification.